+++ partiql-eval/src/plan.rs
@@ -1,13 +1,13 @@
 use itertools::{Either, Itertools};
+use ordered_float::OrderedFloat;
 use partiql_logical as logical;
-use petgraph::prelude::StableGraph;
-use std::collections::HashMap;
-
 use partiql_logical::{
     AggFunc, BagOperator, BinaryOp, BindingsOp, CallName, GroupingStrategy, IsTypeExpr, JoinKind,
-    LogicalPlan, OpId, PathComponent, Pattern, PatternMatchExpr, SearchedCase, SetQuantifier,
+    Lit, LogicalPlan, OpId, PathComponent, Pattern, PatternMatchExpr, SearchedCase, SetQuantifier,
     SortSpecNullOrder, SortSpecOrder, Type, UnaryOp, ValueExpr, VarRefType,
 };
+use petgraph::prelude::StableGraph;
+use std::collections::HashMap;
 
 use crate::error::{ErrorNode, PlanErr, PlanningError};
 use crate::eval;
@@ -26,6 +26,7 @@ use crate::eval::expr::{
 use crate::eval::EvalPlan;
 use partiql_catalog::catalog::{Catalog, FunctionEntryFunction};
 use partiql_value::Value::Null;
+use partiql_value::{Bag, List, Tuple, Value};
 
 #[macro_export]
 macro_rules! correct_num_args_or_err {
@@ -392,7 +393,10 @@ impl<'c> EvaluatorPlanner<'c> {
             ),
             ValueExpr::Lit(lit) => (
                 "literal",
-                EvalLitExpr { lit: *lit.clone() }.bind::<{ STRICT }>(vec![]),
+                EvalLitExpr {
+                    lit: Value::from(lit.as_ref().clone()),
+                }
+                .bind::<{ STRICT }>(vec![]),
             ),
             ValueExpr::Path(expr, components) => (
                 "path",
@@ -565,7 +569,7 @@ impl<'c> EvaluatorPlanner<'c> {
                             n.lhs.clone(),
                             n.rhs.clone(),
                         )),
-                        Box::new(ValueExpr::Lit(Box::new(Null))),
+                        Box::new(ValueExpr::Lit(Box::new(logical::Lit::Null))),
                     )],
                     default: Some(n.lhs.clone()),
                 });
@@ -794,7 +798,7 @@ mod tests {
         // report the error.
         let mut logical = LogicalPlan::new();
         fn lit_int(i: usize) -> ValueExpr {
-            ValueExpr::Lit(Box::new(Value::from(i)))
+            ValueExpr::Lit(Box::new(logical::Lit::Int64(i as i64)))
         }
 
         let expq = logical.add_operator(BindingsOp::ExprQuery(ExprQuery {